Add verbatim handler.
Verbatim allows you to require and include arbitrary racket code
that gets copied to the module, you guessed it, verbatim.

Well, not quite. It wraps it with a `(begin ...)` so that `require`
forms will work appropriately. This is all a big hack, but you
should also get error messages for when reading the code fails.